fix: clear ebpf map after processing (#6)
Currently the main loop does a `shouldInclude` check on each entry
returned by `processMap`. This check compares some of the event's
details as well as a timestamp to see if it should be printed out to
stdout. However, `processMap` is setting this to `time.Now` on each
call, so no event will ever satisfy the `shouldInclude` check.

This results in every event being outputted to stdout each time the main
loop runs.

This change clears the bpf.PktStats map at the end of each `processMap`
run. This causes only new events since the last run of `processMap` to
be output to stdout.